Adding Read functionality to the Benchmark suite
* directory listing completed, (only list files in a directory and not recursive )
* read and write callbacks defined and depending upon io_type, appropriate callback is defined
* Read callback function has been completed ( first cuts )
* Reading works in multiple threads fine

+++ main.cc
@@ -76,7 +76,7 @@ int main(int argc, char **argv) {
                         blocks_count = atoi(optarg);
                         break;
                     case 5:
-                        block_size = atof(optarg);
+                        block_size = atoi(optarg);
                         break;
 
                     default:
